+++ src/Mammoth.Extensions.DependencyInjection/ServiceProviderExtensions.Reflection.cs
@@ -1,151 +1,91 @@
-﻿// deprecated
-
-using Microsoft.Extensions.DependencyInjection;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Reflection;
+﻿using System.Reflection;
 
 namespace Mammoth.Extensions.DependencyInjection
 {
-	// Instead of using reflection to access internal properties of the service provider, consider using the following approach:
-	// Resolve all services: another approach to consider is "enumerate all keys": https://github.com/dotnet/runtime/issues/91466#issuecomment-1723532096
-	// We can create a custom factory provider that will register some structure with all the
-	// information needed to know what services are registered (like service type, service key, etc).
-	// we can then use these structures to resolve all services.
-	// This approach is more reliable and less likely to break in future versions of the framework.
-
-	// These extension methods were disabled in favor of those that use the ServiceProviderFactory approach.
-
 	/// <summary>
-	/// Provides extension methods for <see cref="IServiceProvider"/>.
+	/// Provides extension methods for IServiceProvider to access internal properties.
+	/// Includes methods to retrieve the IsScope property and Disposables collection.
 	/// </summary>
-	public static class ServiceProviderExtensions
+	public static partial class ServiceProviderExtensions
 	{
 		/// <summary>
-		/// Determines whether the specified service type is registered in the service provider (keyed or non-keyed).
+		/// Retrieves the service provider from a given instance, potentially accessing a private 'Root' property.
+		/// This is a very fragile way to obtain internal service provider implementation detail for testing purposes.
+		/// Look at the internal ServiceProvider implementation:
+		/// https://github.com/dotnet/runtime/blob/main/src/libraries/Microsoft.Extensions.DependencyInjection/src/ServiceProvider.cs
 		/// </summary>
-		/// <remarks>
-		/// WARNING: it uses reflection to access internal properties of the service provider, so it may break in future versions of the framework.
-		/// </remarks>
-		public static bool IsServiceRegistered(this IServiceProvider serviceProvider, Type serviceType)
+		/// <returns>Returns the service provider, which may be the original or the root service provider if accessible.</returns>
+		private static IServiceProvider GetServiceProviderEngineScope(IServiceProvider serviceProvider)
 		{
-			if (serviceType is null)
+			var sp = serviceProvider;
+			var rootProperty = serviceProvider.GetType().GetProperty("Root",
+				BindingFlags.NonPublic | BindingFlags.Instance);
+			if (rootProperty != null)
 			{
-				throw new ArgumentNullException(nameof(serviceType));
+				sp = (IServiceProvider)rootProperty.GetValue(serviceProvider)!;
 			}
-
-			var descriptors = GetServiceDescriptors(serviceProvider);
-			return descriptors?.Any(sd => sd.ServiceType == serviceType) ?? false;
+			return sp;
 		}
 
 		/// <summary>
-		/// Determines whether the specified service type is registered in the service provider (keyed or non-keyed).
+		/// Gets the value of the internal IsScope property from a ServiceProvider
 		/// </summary>
-		/// <remarks>
-		/// WARNING: it uses reflection to access internal properties of the service provider, so it may break in future versions of the framework.
-		/// </remarks>
-		public static bool IsServiceRegistered<TService>(this IServiceProvider serviceProvider)
+		internal static bool GetIsRootScope(this IServiceProvider serviceProvider)
 		{
-			return IsServiceRegistered(serviceProvider, typeof(TService));
+			// if it has a Root "property" that will be the root scope (an instance of ServiceProviderEngineScope)
+			// otherwise it's already an instance of ServiceProviderEngineScope and it's a scope.
+			var sp = GetServiceProviderEngineScope(serviceProvider);
+
+			// Get the IsScope property using reflection
+			var isScopeProperty = sp.GetType().GetProperty("IsRootScope",
+				BindingFlags.Public | BindingFlags.NonPublic | BindingFlags.Instance);
+
+			if (isScopeProperty != null)
+			{
+				return (bool)isScopeProperty.GetValue(sp);
+			}
+
+			throw new InvalidOperationException("Internal implementation of ServiceProvider changed: cannot access IsRootScope property.");
 		}
 
 		/// <summary>
-		/// Determines whether the specified service type is registered in the service provider (keyed or non-keyed).
+		/// Gets the internal Disposables collection from a ServiceProvider
 		/// </summary>
-		/// <remarks>
-		/// WARNING: it uses reflection to access internal properties of the service provider, so it may break in future versions of the framework.
-		/// </remarks>
-		public static bool IsServiceRegistered(this IServiceProvider serviceProvider, object serviceKey)
+		internal static IEnumerable<object> GetDisposables(this IServiceProvider serviceProvider)
 		{
-			if (serviceKey is null)
-			{
-				throw new ArgumentNullException(nameof(serviceKey));
-			}
+			// if it has a Root "property" that will be the root scope (an instance of ServiceProviderEngineScope)
+			// otherwise it's already an instance of ServiceProviderEngineScope and it's a scope.
+			var sp = GetServiceProviderEngineScope(serviceProvider);
 
-			var descriptors = GetServiceDescriptors(serviceProvider);
-			return descriptors?.Any(sd => sd.ServiceKey?.Equals(serviceKey) == true) ?? false;
-		}
+			/*
+			// Get the Disposables field using reflection
+			var disposablesField = sp.GetType().GetField("_disposables",
+				BindingFlags.NonPublic | BindingFlags.Instance);
 
-		/// <summary>
-		/// Resolves all the services of the specified ServiceTye (both keyed and non-keyed) from the service provider.
-		/// </summary>
-		/// <remarks>
-		/// <para>WARNING: it uses reflection to access internal properties of the service provider, so it may break in future versions of the framework.</para>
-		/// <para>WARNING: It might be removed once enumerating all services with KeyedService.AnyKey work as expected.</para>
-		/// </remarks>
-		public static IEnumerable<object> GetAllServices(this IServiceProvider serviceProvider, Type serviceType)
-		{
-			var descriptors = GetServiceDescriptors(serviceProvider);
-			if (descriptors == null)
+			if (disposablesField != null)
 			{
-				return Enumerable.Empty<object>();
-			}
-			// find out all the unique services keyed names
-			var serviceKeys = descriptors
-				.Where(sd => sd.ServiceType == serviceType)
-				.Select(sd => sd.ServiceKey)
-				.Distinct()
-				.ToList();
-			var serviceList = new List<object>();
-			// null key to get all non-keyed services
-			foreach (var serviceKey in serviceKeys)
-			{
-				var services = serviceProvider.GetKeyedServices(serviceType, serviceKey);
-				if (services?.Any() == true)
+				var disposables = disposablesField.GetValue(sp);
+				if (disposables is IEnumerable<object> disposablesList)
 				{
-					serviceList.AddRange(services!);
+					return disposablesList;
 				}
 			}
-			return serviceList;
-		}
+			*/
 
-		/// <summary>
-		/// Resolves all the services of the specified ServiceTye (both keyed and non-keyed) from the service provider.
-		/// </summary>
-		/// <remarks>
-		/// <para>WARNING: it uses reflection to access internal properties of the service provider, so it may break in future versions of the framework.</para>
-		/// <para>WARNING: It might be removed once enumerating all services with KeyedService.AnyKey work as expected.</para>
-		/// </remarks>
-		public static IEnumerable<TService> GetAllServices<TService>(this IServiceProvider serviceProvider)
-		{
-			var descriptors = GetServiceDescriptors(serviceProvider);
-			if (descriptors == null)
+			// Check alternative field names (since internal implementation might vary)
+			var disposablesField = sp.GetType().GetProperty("Disposables",
+				BindingFlags.NonPublic | BindingFlags.Instance);
+
+			if (disposablesField != null)
 			{
-				return Enumerable.Empty<TService>();
-			}
-			// find out all the unique services keyed names
-			var serviceKeys = descriptors
-				.Where(sd => sd.ServiceType == typeof(TService))
-				.Select(sd => sd.ServiceKey)
-				.Distinct()
-				.ToList();
-			var serviceList = new List<TService>();
-			// null key to get all non-keyed services
-			foreach (var serviceKey in serviceKeys)
-			{
-				var services = serviceProvider.GetKeyedServices<TService>(serviceKey);
-				if (services?.Any() == true)
+				var disposables = disposablesField.GetValue(sp);
+				if (disposables is IEnumerable<object> disposablesList)
 				{
-					serviceList.AddRange(services!);
+					return disposablesList;
 				}
 			}
-			return serviceList;
-		}
 
-		private static IEnumerable<ServiceDescriptor> GetServiceDescriptors(IServiceProvider serviceProvider)
-		{
-			var callSiteFactoryProperty = serviceProvider.GetType().GetProperty("CallSiteFactory", BindingFlags.NonPublic | BindingFlags.Instance);
-			if (callSiteFactoryProperty != null)
-			{
-				var callSiteFactory = callSiteFactoryProperty.GetValue(serviceProvider);
-				var serviceDescriptorsProperty = callSiteFactory?.GetType().GetProperty("Descriptors", BindingFlags.NonPublic | BindingFlags.Instance);
-				if (serviceDescriptorsProperty != null)
-				{
-					return (IEnumerable<ServiceDescriptor>)serviceDescriptorsProperty.GetValue(callSiteFactory);
-				}
-			}
-			throw new InvalidOperationException("Internal implementation of ServiceProvider changed, cannot enumerate service descriptors.");
+			throw new InvalidOperationException("Internal implementation of ServiceProvider changed: cannot access Disposable property.");
 		}
 	}
 }