+++ Changelog.md
@@ -2,6 +2,8 @@
 
 ## vNext
 
+- Detect incorrect usage of Transient Disposable services: resolving a transient disposable result in a memory leak [#5](https://github.com/PrimordialCode/Mammoth.Extensions.DependencyInjection/issues/5).
+
 ## 0.4.0
 
 - Added ServiceProvider extension methods to check if Registered Services are: Transient, Scoped, Singleton [#1](https://github.com/PrimordialCode/Mammoth.Extensions.DependencyInjection/issues/1)